Adds distinction between content/distribution digests
Being pkg/drivers/kic/types.go the source of truth for the version of
the container we're using to instantiate our kübernetes cluster in,
it should be appropriate to hardcode here the
imageId(a.k.a. contentDigest) so that it could be later used
as a discriminant to invalidate minikube's cache

contentDigest is the most reliable way to address image content:
if the image is tampered with after push to a registry,
the contentDigest we'd see after pull,
would be different than the one hardcoded here.
It is also part of the image itself, i.e. part of the tar archive;
thus giving us a way to always know if the cache is up to date,
even offline.

distributionDigest is the most reliable way to determine which
image we're looking to pull from a registry; a tag can be detached
from an image and recycled, referencing another one, with different
content.
It is not part of the image itself; it is computed on the image in
compressed state.. and since different engines/mechanisms could use
different types of compression, this digest is totally unreliable
as a way to address content.

[*] refs:
https://windsock.io/explaining-docker-image-ids/
https://github.com/google/go-containerregistry/issues/895#issuecomment-753521526
https://stackoverflow.com/questions/45533005/why-digests-are-different-depend-on-registry
https://blog.aquasec.com/docker-image-tags -- follow links