Changes cache-invalidation mechanism to depend on contentDigest
The idea is that the distributionDigest was not
meant to be used as something content-addressable.
So instead of invalidating the kicBase cache based on the
distributionDigest present in the kicDriver's storage,
we could move to something like this:

.User expresses preferences in regard of kicBase image,
 by specifying the image by name:tag@digest, name:tag,... whatever
.The image's specified name is sanitized
 and used as filename for the tarbal in minikube's cache[*]
.Image stored inside kicDriver(docker, podman,...) is validated
 against the image's contentDigest(imageID), which we're retrieving
 directly from the .tar archive.. which we're selecting from the
 cache, by sanitized name

I initially tried implementing a mechanism based on a json file that
would contain entries in regard of cached images, plus other infos
that we would need to address content, like
distributionDigest(registry digest)...
Something like a repositories.json file.. in a docker fashion
(/var/lib/docker/image/overlay2/repositories.json)
That would lead to complications like
"what if user removes a .tar to save space?",
(basically.. how we keep the file up-to-date with cache content)
that would add extra complexity in solving.. like
some logic that would be called on startup, that would read all
cached files and reflect in the file.

As far as I can understand our usecase, which doesn't seem very
complex(I might be wrong..), some kind of simple/raw mechanism like
the one in this commit could suffice..
Like we don't have to maintain a lot of images..
The source of truth is hardcoded in the sources..
the image/file relation for the kicBase cache
is based on the sanitize(img) mechanism..
...

[+] refs..
https://windsock.io/explaining-docker-image-ids/
https://blog.aquasec.com/docker-image-tags
https://stackoverflow.com/questions/45533005/why-digests-are-different-depend-on-registry

[*] TODO: this could be one possible shortcoming..
  - what if user wants the image:tag to be always up-to-date with
    registry? perhaps some flag?