Corrects typo in begindownloadkicbaseimage() imageLoad errcheck
We're checking if there is no error.. so that we could inform the user
that we've successfully loaded the image.
Otherwise we would fail to load image(even tho everything went fine..)
and skip to the following fallback image, to then find it already
inside the driver, skipping the load (both images got same contentDigest).